* New Features
  - Add compatibility with **Unity 2020.1**
  - Add support for **Unity XR Platform** (OpenVR, Oculus, and Windows MR)
    - VIU Settings automatically installs proper XR Loader from **PackageManager** for supported devices
      - **OpenVR XR Plugin**
      - **Oculus XR Plugin**
      - **Windows XR Plugin**
  - Add support for Oculus controller render model (Requires [Oculus SDK](https://assetstore.unity.com/packages/tools/integration/oculus-integration-82022))
  - Add new **ControllerButton.DPadCenter** & **ControllerButton.DPadCenterTouch**
    - Also change **VIUSettings.virtualDPadDeadZone** default value from 0.15 to **0.25**
    - This settings is not available in VIU Settings UI yet, but can be manually modified in HTC.UnityPlugin/ViveInputUtility/Resources/**VIUSettings.asset**

* Changes
  - Now use relative path when choosing Oculus Android AndroidManifest.xml file with picker (#175)
  - Update **ControllerManagerSample**
    - Fix support for StickyGrabbables
    - Clean up side cases where updateactivity was needed
    - Add more button options for laser pointer
    - Fix typo
  - Slightly change **ColliderEventCaster**'s behaviour
    - Now IColliderEventPressUpHandler will be treggered only if IColliderEventPressDownHandler also implemented
    - This is for aligning IPointerUpHandler and IPointerDownHandler behaviour
  - Improve **Grabbable**
    - Now **BasicGrabbable** & **StickyGrabbable** able to accept more then one grab button
      - Add new property **primaryGrabButton** so it can be specified with VIU ControllerButton
      - Obsolete property grabButton and add new property **secondaryGrabButton** as a substitute
      - Note that you must setup **ViveColliderEventCaster** properly to send the specified grab button event
  - Improve **Teleportable**
    - Now **Teleportable** able to accept more then one teleport button
      - Add new property **PrimeryTeleportButton** so it can specify with VIU ControllerButton
        - Obsolete property teleportButton and add new property **SecondaryTeleportButton** as a substitute
        - Note that you must setup **ViveRaycaster** properly to send the specified teleport button event
      - Add new property **TriggeredType**
        - **ButtonUp** : perform teleport on button press up (default)
        - **ButtonDown** : perform teleport on button press down
        - **ButtonClick** : perform teleport on button press up only if pointed object when press down/up are the same
      - Add new property **RotateToHitObjectFront**
        - When set to true, teleportation will rotate pivot front to hit object front
      - Add new property **TeleportToHitObjectPivot**
        - When set to true, teleportation will move pivot to hit object pivot instead of the hit point
      - Add new property **UseSteamVRFade**
        - Only works when [SteamVR Plugin](https://assetstore.unity.com/packages/tools/integration/steamvr-plugin-32647) is installed
        - When set to false or SteamVR Plugin is not installed, the teleportation will delay for half of fadeDuration without fading effect
        - This provides an option for developer to implement their custom fading effect in the OnBeforeTeleport event
      - Add event **OnBeforeTeleport(Teleportable src, RaycastResult hitResult, float delay)**
        - Emit before fade duration start counting down
        - Usually delay argument is half of fade duration (0 if fadeDuration is ignored)
        - Possible usage is to start custom fading effect in this callback
      - Add event **OnAfterTeleport(Teleportable src, RaycastResult hitResult, float delay)**
        - Emit after teleportation is performed
      - Add static event **OnBeforeAnyTeleport(Teleportable src, RaycastResult hitResult, float delay)**
        - Static version of OnBeforeTeleport
        - Emit before OnBeforeTeleport
      - Add static event **OnAfterAnyTeleport(Teleportable src, RaycastResult hitResult, float delay)**
        - Static version of OnAfterTeleport
        - Emit before OnAfterTeleport
      - Add property **AdditionalTeleportRotation**
        - The rotation value will be multiplied on the target (around pivot) when teleporting
        - Possible usage is to set the value (according to other input like pad or joystick direction) in the OnBeforeTeleport callback
      - Add method **AbortTeleport()**
        - Cancel the teleportation during fading
        - Possible usage is to abort in the OnBeforeTeleport callback to perform a custom teleportation
        - Another usage is to interrupt the fade-in effect progress so that the Teleportable able to trigger next teleport event immediatly

* Bug Fixes
  - Fix left Cosmos controller did not bind button X and Y
  - Fix saving of Oculus Android XML path setting (#175)
  - Fix applying some recommended settings didn't trigger editor to compile
  - Fix "recommended settings" button in VIU Settings disappeared after exiting editor play mode

+++ Assets/HTC.UnityPlugin/ViveInputUtility/Scripts/Editor/VRPlatformSettings/OculusGoSettings.cs
@@ -1,4 +1,4 @@
-﻿//========= Copyright 2016-2019, HTC Corporation. All rights reserved. ===========
+﻿//========= Copyright 2016-2020, HTC Corporation. All rights reserved. ===========
 
 using HTC.UnityPlugin.VRModuleManagement;
 using UnityEditor;
@@ -415,6 +415,7 @@ namespace HTC.UnityPlugin.Vive
     }
     public static partial class VIUSettingsEditor
     {
+        public const string URL_OCULUS_VR_PLUGIN = "https://www.assetstore.unity3d.com/en/#!/content/82022";
         private const string OCULUS_ANDROID_PACKAGE_NAME = "com.unity.xr.oculus.android";
 
         public static bool canSupportOculusGo
@@ -466,8 +467,10 @@ namespace HTC.UnityPlugin.Vive
             {
                 get
                 {
-#if UNITY_2018_1_OR_NEWER
-                    return activeBuildTargetGroup == BuildTargetGroup.Android && PackageManagerHelper.IsPackageInList(OCULUS_ANDROID_PACKAGE_NAME);
+#if UNITY_2020_1_OR_NEWER
+                    return activeBuildTargetGroup == BuildTargetGroup.Android && PackageManagerHelper.IsPackageInList(OCULUS_XR_PACKAGE_NAME);
+#elif UNITY_2018_1_OR_NEWER
+                    return activeBuildTargetGroup == BuildTargetGroup.Android && (PackageManagerHelper.IsPackageInList(OCULUS_ANDROID_PACKAGE_NAME) || PackageManagerHelper.IsPackageInList(OCULUS_XR_PACKAGE_NAME));
 #elif UNITY_5_6_OR_NEWER
                     return activeBuildTargetGroup == BuildTargetGroup.Android && VRModule.isOculusVRPluginDetected;
 #else
@@ -482,11 +485,17 @@ namespace HTC.UnityPlugin.Vive
                 get
                 {
                     if (!canSupport) { return false; }
-                    if (!VIUSettings.activateOculusVRModule) { return false; }
-                    if (!OculusSDK.enabled) { return false; }
                     if (PlayerSettings.Android.minSdkVersion < AndroidSdkVersions.AndroidApiLevel21) { return false; }
                     if (PlayerSettings.graphicsJobs) { return false; }
                     if ((PlayerSettings.colorSpace == ColorSpace.Linear || PlayerSettings.gpuSkinning) && !GraphicsAPIContainsOnly(BuildTarget.Android, GraphicsDeviceType.OpenGLES3)) { return false; }
+
+#if UNITY_2019_3_OR_NEWER
+                    if (!((VIUSettings.activateOculusVRModule && OculusSDK.enabled)
+                          || (VIUSettings.activateUnityXRModule && XRPluginManagementUtils.IsXRLoaderEnabled(OculusVRModule.OCULUS_XR_LOADER_NAME, requirdPlatform)))) { return false; }
+#else
+                    if (!VIUSettings.activateOculusVRModule) { return false; }
+                    if (!OculusSDK.enabled) { return false; }
+#endif
                     return true;
                 }
                 set
@@ -512,8 +521,19 @@ namespace HTC.UnityPlugin.Vive
                     }
 
                     VIUSettings.activateOculusVRModule = value;
+#if UNITY_2020_1_OR_NEWER
+                    XRPluginManagementUtils.SetXRLoaderEnabled(OculusVRModule.OCULUS_XR_LOADER_CLASS_NAME, requirdPlatform, value);
+                    OculusSDK.enabled = value && !PackageManagerHelper.IsPackageInList(OCULUS_XR_PACKAGE_NAME);
+                    VIUSettings.activateUnityXRModule = XRPluginManagementUtils.IsAnyXRLoaderEnabled(requirdPlatform);
+#elif UNITY_2019_3_OR_NEWER
+                    XRPluginManagementUtils.SetXRLoaderEnabled(OculusVRModule.OCULUS_XR_LOADER_CLASS_NAME, requirdPlatform, value);
+                    OculusSDK.enabled = value && !PackageManagerHelper.IsPackageInList(OCULUS_XR_PACKAGE_NAME);
+                    VIUSettings.activateUnityXRModule = XRPluginManagementUtils.IsAnyXRLoaderEnabled(requirdPlatform);
+                    VIUSettings.activateUnityNativeVRModule = value || supportOpenVR;
+#else
                     OculusSDK.enabled = value;
                     VIUSettings.activateUnityNativeVRModule = value;
+#endif
                 }
 #else
                 get { return false; }
@@ -543,6 +563,17 @@ namespace HTC.UnityPlugin.Vive
                         GUILayout.FlexibleSpace();
                         ShowSwitchPlatformButton(BuildTargetGroup.Android, BuildTarget.Android);
                     }
+#if UNITY_2019_3_OR_NEWER
+                    else if (!PackageManagerHelper.IsPackageInList(OCULUS_XR_PACKAGE_NAME))
+                    {
+                        GUI.enabled = false;
+                        ShowToggle(new GUIContent(title, "Oculus XR Plugin package required."), false, GUILayout.Width(230f));
+                        GUI.enabled = true;
+                        GUILayout.FlexibleSpace();
+                        ShowAddPackageButton("Oculus XR Plugin", OCULUS_XR_PACKAGE_NAME);
+                    }
+#endif
+#if !UNITY_2020_1_OR_NEWER
                     else if (!PackageManagerHelper.IsPackageInList(OCULUS_ANDROID_PACKAGE_NAME))
                     {
                         GUI.enabled = false;
@@ -551,6 +582,7 @@ namespace HTC.UnityPlugin.Vive
                         GUILayout.FlexibleSpace();
                         ShowAddPackageButton("Oculus Android", OCULUS_ANDROID_PACKAGE_NAME);
                     }
+#endif
                     else if (!VRModule.isOculusVRPluginDetected)
                     {
                         GUI.enabled = false;
@@ -576,10 +608,18 @@ namespace HTC.UnityPlugin.Vive
                                                 VIUSettings.oculusVRAndroidManifestPath, style);
                         if (GUILayout.Button("Open", new GUILayoutOption[] { GUILayout.Width(44), GUILayout.Height(18) }))
                         {
-                            VIUSettings.oculusVRAndroidManifestPath = EditorUtility.OpenFilePanel("Select AndroidManifest.xml", string.Empty, "xml");
+                            var path = EditorUtility.OpenFilePanel("Select AndroidManifest.xml", string.Empty, "xml");
+                            if (path.Length != 0)
+                            {
+                                // make relative path if it is under Assets folder.
+                                if (path.StartsWith(Application.dataPath))
+                                {
+                                    path = "Assets" + path.Substring(Application.dataPath.Length);
+                                }
+                                VIUSettings.oculusVRAndroidManifestPath = path;
+                            }
                         }
 
-                        EditorGUI.BeginChangeCheck();
                         EditorGUILayout.EndHorizontal();
 
                         EditorGUILayout.BeginHorizontal();
@@ -593,7 +633,6 @@ namespace HTC.UnityPlugin.Vive
                             EditorGUILayout.HelpBox("File does not existed!", MessageType.Warning);
                         }
 
-                        EditorGUI.BeginChangeCheck();
                         EditorGUILayout.EndHorizontal();
                         EditorGUI.indentLevel -= 2;
                     }
@@ -622,11 +661,27 @@ namespace HTC.UnityPlugin.Vive
 
                 if (File.Exists(VIUSettings.oculusVRAndroidManifestPath))
                 {
-                    File.Copy(VIUSettings.oculusVRAndroidManifestPath, "Assets/Plugins/Android/AndroidManifest.xml", true);
+                    if (Directory.Exists("Assets/Plugins/Android/AndroidManifest.xml"))
+                    {
+                        File.Copy(VIUSettings.oculusVRAndroidManifestPath, "Assets/Plugins/Android/AndroidManifest.xml", true);
+                    }
+                    else
+                    {
+                        Directory.CreateDirectory("Assets/Plugins/Android/");
+                        File.Copy(VIUSettings.oculusVRAndroidManifestPath, "Assets/Plugins/Android/AndroidManifest.xml", true);
+                    }
                 }
                 else if (File.Exists(defaultAndroidManifestPath))
                 {
-                    File.Copy(defaultAndroidManifestPath, "Assets/Plugins/Android/AndroidManifest.xml", true);
+                    if (Directory.Exists("Assets/Plugins/Android/AndroidManifest.xml"))
+                    {
+                        File.Copy(defaultAndroidManifestPath, "Assets/Plugins/Android/AndroidManifest.xml", true);
+                    }
+                    else
+                    {
+                        Directory.CreateDirectory("Assets/Plugins/Android/");
+                        File.Copy(defaultAndroidManifestPath, "Assets/Plugins/Android/AndroidManifest.xml", true);
+                    }
                 }
             }
 #endif
